+++ server/src/main/java/com/inventage/portal/gateway/proxy/middleware/proxy/ProxyMiddleware.java
@@ -3,7 +3,6 @@ package com.inventage.portal.gateway.proxy.middleware.proxy;
 import com.inventage.portal.gateway.proxy.middleware.Middleware;
 import com.inventage.portal.gateway.proxy.middleware.TraceMiddleware;
 import com.inventage.portal.gateway.proxy.middleware.proxy.contextAware.ContextAwareHttpServerRequest;
-import io.netty.handler.codec.http.HttpHeaderNames;
 import io.opentelemetry.api.trace.Span;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
@@ -11,6 +10,9 @@ import io.vertx.core.MultiMap;
 import io.vertx.core.Vertx;
 import io.vertx.core.http.HttpClient;
 import io.vertx.core.http.HttpClientOptions;
+import io.vertx.core.http.HttpHeaders;
+import io.vertx.core.http.HttpServerRequest;
+import io.vertx.core.net.HostAndPort;
 import io.vertx.core.net.JksOptions;
 import io.vertx.ext.web.RoutingContext;
 import io.vertx.httpproxy.HttpProxy;
@@ -30,9 +32,8 @@ public class ProxyMiddleware extends TraceMiddleware {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ProxyMiddleware.class);
 
-    private static final String X_FORWARDED_PROTO = "X-Forwarded-Proto";
-    private static final String X_FORWARDED_HOST = "X-Forwarded-Host";
-    private static final String X_FORWARDED_PORT = "X-Forwarded-Port";
+    private static final CharSequence X_FORWARDED_PROTO = HttpHeaders.createOptimized("x-forwarded-proto");
+    private static final CharSequence X_FORWARDED_PORT = HttpHeaders.createOptimized("x-forwarded-port");
 
     private static final boolean DEFAULT_HTTPS_TRUST_ALL = true;
     private static final boolean DEFAULT_HTTPS_VERIFY_HOSTNAME = false;
@@ -64,6 +65,9 @@ public class ProxyMiddleware extends TraceMiddleware {
         this.serverHost = serverHost;
         this.serverPort = serverPort;
 
+        lowercaseHostHeader(httpProxy);
+        setHostHeader(httpProxy);
+        setXForwardedHeaders(httpProxy);
         applyModifiers(httpProxy);
     }
 
@@ -90,20 +94,96 @@ public class ProxyMiddleware extends TraceMiddleware {
     public void handleWithTraceSpan(RoutingContext ctx, Span span) {
         LOGGER.debug("{}: Handling '{}'", name, ctx.request().absoluteURI());
 
-        useOrSetHeader(X_FORWARDED_PROTO, ctx.request().scheme(), ctx.request().headers());
-        if (ctx.request().authority() != null) {
-            useOrSetHeader(X_FORWARDED_HOST, ctx.request().authority().host(), ctx.request().headers());
-            useOrSetHeader(X_FORWARDED_PORT, String.valueOf(ctx.request().authority().port()), ctx.request().headers());
-        }
-
-        ctx.request().headers().set(HttpHeaderNames.HOST, serverHost);
-        LOGGER.debug("'{}' is sending to '{}:{}{}'", name, serverHost, serverPort, ctx.request().uri());
+        LOGGER.debug("'{}' is proxying to '{}:{}{}'", name, serverHost, serverPort, ctx.request().uri());
         try {
             httpProxy.handle(new ContextAwareHttpServerRequest(ctx.request(), ctx));
         } catch (Exception e) {
             LOGGER.error("Error while proxying request", e);
             ctx.fail(e);
         }
+    }
+
+    /**
+     * Due to a bug in the vertx-http-proxy comparing the host header case-sensitive with "host",
+     * we have to make sure the incoming request host header is lowercase as well i.e. not "Host".
+     * Otherwise, we have a weird combination of having a host header value in the request.headers()
+     * and another host header value in the request.host.
+     * Apparently, the host header value in request.headers() has presendence.
+     * 
+     * The bug was already fixed and is probably available in the next release.
+     * This interceptor can then be removed.
+     * See: https://github.com/eclipse-vertx/vertx-http-proxy/issues/77
+     * 
+     * @param proxy
+     */
+    protected void lowercaseHostHeader(HttpProxy proxy) {
+        proxy.addInterceptor(new ProxyInterceptor() {
+            @Override
+            public Future<ProxyResponse> handleProxyRequest(ProxyContext proxyContext) {
+                final ProxyRequest incomingRequest = proxyContext.request();
+                if (incomingRequest.headers().contains("Host")) {
+                    LOGGER.debug("lowercasing the host header name");
+                    final String host = incomingRequest.headers().get("Host");
+                    incomingRequest.headers().remove("Host");
+                    incomingRequest.headers().set("host", host);
+                }
+                return proxyContext.sendRequest();
+            }
+        });
+    }
+
+    /**
+     * 
+     * Technically, vertx-http-proxy should set the host header to the target server.
+     * However, due to a bug, it is set to a wrong value.
+     * 
+     * Once this issue has been fixed, this interceptor can be removed.
+     * See: https://github.com/eclipse-vertx/vertx-http-proxy/issues/85
+     * 
+     * @param proxy
+     */
+    protected void setHostHeader(HttpProxy proxy) {
+        proxy.addInterceptor(new ProxyInterceptor() {
+            @Override
+            public Future<ProxyResponse> handleProxyRequest(ProxyContext proxyContext) {
+                final ProxyRequest incomingRequest = proxyContext.request();
+                LOGGER.debug("setting the host header to '{}:{}'", serverHost, serverPort);
+                incomingRequest.setAuthority(HostAndPort.create(serverHost, serverPort));
+                return proxyContext.sendRequest();
+            }
+        });
+    }
+
+    /**
+     * Since version 4.3.5, the vertx-http-proxy sets the 'x-forwarded-host', in case it detects that
+     * the outgoing request has a different host header value than the incoming request.
+     * We previously set that header outself, alongside with 'x-forwarded-proto' and 'x-forwarded-port',
+     * so we still need to set them here.
+     * 
+     * @param proxy
+     */
+    protected void setXForwardedHeaders(HttpProxy proxy) {
+        proxy.addInterceptor(new ProxyInterceptor() {
+            @Override
+            public Future<ProxyResponse> handleProxyRequest(ProxyContext proxyContext) {
+                final ProxyRequest incomingRequest = proxyContext.request();
+                final HttpServerRequest request = incomingRequest.proxiedRequest();
+
+                final String proto = request.scheme();
+                LOGGER.debug("setting the '{}' header to '{}'", X_FORWARDED_PROTO, proto);
+                incomingRequest.putHeader(X_FORWARDED_PROTO, proto);
+
+                if (request.authority() != null) {
+                    final int port = request.authority().port();
+                    if (port > 0) {
+                        LOGGER.debug("setting the '{}' header to '{}'", X_FORWARDED_PORT, port);
+                        incomingRequest.putHeader(X_FORWARDED_PORT, String.valueOf(port));
+                    }
+                }
+
+                return proxyContext.sendRequest();
+            }
+        });
     }
 
     protected void applyModifiers(HttpProxy proxy) {
@@ -162,35 +242,5 @@ public class ProxyMiddleware extends TraceMiddleware {
                 return proxyContext.sendResponse();
             }
         });
-    }
-
-    /**
-     * If the given header name is already contained in the request, this header will be used, otherwise the given header value is used.
-     *
-     * @param headerName
-     *            to check the request for
-     * @param headerValue
-     *            to use if the header name is not yet in the request
-     * @param headers
-     *            of the request
-     */
-    protected void useOrSetHeader(String headerName, String headerValue, MultiMap headers) {
-        if (headers.contains(headerName)) { // use
-            LOGGER.debug("Using provided header '{}' with '{}'", headerName, headers.get(headerName));
-        } else { // set
-            headers.add(headerName, headerValue);
-            LOGGER.debug("Set header '{}' to '{}'", headerName, headers.get(headerName));
-        }
-    }
-
-    protected void addOrSetHeader(String headerName, String headerValue, MultiMap headers) {
-        if (headers.contains(headerName)) { // add == append
-            final String existingHeader = headers.get(headerName);
-            headers.set(headerName, existingHeader + ", " + headerValue);
-            LOGGER.debug("Appended to header '{}' to '{}' ", headerName, headers.get(headerName));
-        } else { // set
-            headers.add(headerName, headerValue);
-            LOGGER.debug("Set header '{}' to '{}'", headerName, headers.get(headerName));
-        }
     }
 }